+++ src/interface.ts
@@ -105,7 +105,7 @@ export interface Critical {
  * 	- Using it, it will roll normally and send the result to the user 
  * - If the user doesn't exists or their stat was deleted: the bot will send a message to inform the user that he is not registered and roll normally, ignoring the statistique/characters (theses will be send into the comments part)
  */
-export interface User {
+export interface UserData {
 	userName?: string; //by default, will be the id of the user, if changed to a string, it will be used
 	stats?: {
 		[name: string] : number;
